+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v3.0.0-rc.1 - 2023-06-21
+## v3.0.0-rc.2 - unreleased
 
 Most notably, v3 switches the letter case of generated IDs from uppercase (e.g.,
 "036Z951MHJIKZIK2GSL81GR7L") to lowercase (e.g., "036z951mhjikzik2gsl81gr7l"),
@@ -20,6 +20,8 @@ a case-insensitive scheme. Other changes include the removal of deprecated APIs.
 ### Changed
 
 - Letter case of generated IDs from uppercase to lowercase
+- `Scru128Generator`'s prerequisite from `rand::RngCore` to `Scru128Rng` to
+  relax hard dependency on `rand` crate
 - Internal representation of `Scru128Id` from `u128` to big-endian byte array
   representation